requirements.txt
# Saudi Arabia Q&A Agent Dependencies

# Core LLM and Agent Framework
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.40
openai>=1.0.0

# Web Search
tavily-python>=0.3.0

# Observability and Evaluation
langfuse>=2.0.0

# CLI and Display
rich>=13.0.0
python-dotenv>=0.19.0

# Data Processing
pydantic>=2.0.0

# Development Dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# isort>=5.10.0